feat: add RankSuitIconLoader module
